Rename ResolverEngineGatherer
